+++ changelog.md
@@ -1,3 +1,9 @@
+0.2.0.0
+-------
+
+- missing instances
+- knowDiscretized -> fromUnknown
+
 0.1.0.0
 -------
 - depending on @semirings@ now, instead of our own type class. breaks existing